requirements.txt
rq==1.16.2
rq-scheduler==0.13.1
redis==5.0.1
requests==2.32.3
pydantic==2.10.6
PyYAML==6.0.2
python-dotenv==1.0.1
Jinja2==3.1.5
tenacity==9.0.0
pytest==8.3.4
cryptography==44.0.0
